\name{add}
\alias{add}
\title{
 Sum(+): Tropical Algebra Function
}
\description{
 It takes two parameters(numeric) in X and y and returns the tropical sum of these numbers e.g. x + y=minimum(x,y).
}
\usage{
add(x,y)
}
\arguments{
  \item{x}{
 A numeric value
}
  \item{y}{
 A numeric value
}
}
\details{
It takes two parameters(numeric) in X and y and returns the tropical sum of these numbers e.g. x + y=minimum(x,y).
}
\value{
  Tropical sum of \code{x} and \code{y}
}
\examples{
add(1,2)
}

