\name{multiply}
\alias{multiply}

\title{
   Product: Tropical Algebra Function
}
\description{
 This function returns the product of X and Y according to tropical algebra.
}
\usage{
multiply(x,y)
}

\arguments{
  \item{x}{
A numeric value
}
\item{y}{
A numeric value
}
}
\details{
e.g.
#'   x<- 6
#'   y<- 5
#'   Now function call : multiply(5,6)
#'   The returned value will be 11.
}
\value{
  Returns the tropical multiplication of \code{x} and \code{y}
}
\examples{
 multiply(6,5)

}
